a little nicer script

+++ ipfsgit.sh
@@ -1,10 +1,16 @@
 #!/bin/sh
 
 dir=`mktemp -d`
+clonelog=`mktemp`
 repo=$dir/tohost
-echo rehosting $1
-git clone --bare $1 $repo
-cd $repo && cp objects/pack/*.pack . && \
+echo Rehosting repo at \'$1\'.
+git clone --bare $1 $repo 2> $clonelog
+if [ $? -ne 0 ]; then
+	cat $clonelog
+	exit -1
+fi
+
+cd $repo && cp objects/pack/*.pack . 2> /dev/null && \
 	git unpack-objects < ./*.pack && \
 	rm ./*.pack
 
@@ -12,4 +18,4 @@ cd $repo && git update-server-info
 
 hash=`ipfs add -q -r $repo | tail -n1`
 echo hosted at http://gateway.ipfs.io/ipfs/$hash
-
+echo try it out: 'git clone http://gateway.ipfs.io/ipfs/$hash'
